feat: update offline videos page

+++ src/extension/script/custom/youtube/pages/offline.ts
@@ -1,16 +1,23 @@
 import Lifecycle from '@ext/common/preload/overlay/components/lifecycle'
 import { buildClass } from '@ext/common/preload/overlay/style/class'
-import { dispatchYTNavigate } from '@ext/custom/youtube/module/core/event'
+import { dispatchYTAction, dispatchYTNavigate } from '@ext/custom/youtube/module/core/event'
 import { encodeEntityKey, EntityType } from '@ext/custom/youtube/proto/entity-key'
 import { getYTLocalEntitiesByType, getYTLocalEntityByKey, YTLocalEntity, YTLocalMediaType } from '@ext/custom/youtube/utils/local'
 import { deleteYTOfflineMedia, exportYTOfflineMediaBundle, exportYTOfflineMediaStream, importYTOfflineMediaBundle } from '@ext/custom/youtube/utils/ytom'
 import { PromiseWithProgress } from '@ext/lib/async'
 import van, { ChildDom, State } from 'vanjs-core'
 
-const { button, div, h1, input, p, table, tbody, td, th, thead, tr } = van.tags
+const { button, div, h1, h4, input, p, table, tbody, td, th, thead, tr } = van.tags
 
-const FLEX_ITEM_GROW_STYLE = 'flex-grow:1'
-const SINGLE_LINE_COLUMN = { style: 'width:.1%;white-space:nowrap;text-align:left!important' }
+const COPY_TEXT_TOOLTIP = 'Click to Copy'
+const STYLE_FLEX_CONTAINER = 'display:flex;flex-direction:row;gap:0.5em'
+const STYLE_FLEX_ITEM_GROW = 'flex-grow:1'
+const STYLE_TEXT_SHRINK = 'width:.1%;white-space:nowrap'
+const COLUMN_PROPS_ACTION_HEAD = { style: STYLE_TEXT_SHRINK }
+const COLUMN_PROPS_ACTION_BODY = { style: STYLE_TEXT_SHRINK, rowSpan: 2 }
+const COLUMN_PROPS_INFO_HEAD = { colSpan: 3 }
+const COLUMN_PROPS_INFO_TITLE = { style: 'cursor:pointer', colSpan: 3, title: COPY_TEXT_TOOLTIP, onclick: copyEventTarget }
+const COLUMN_PROPS_INFO_SMALL = { style: 'cursor:pointer;width:25%;white-space:nowrap', title: COPY_TEXT_TOOLTIP, onclick: copyEventTarget }
 
 const enum ExportFormat {
   BUNDLE,
@@ -18,29 +25,51 @@ const enum ExportFormat {
   VIDEO_STREAM
 }
 
-type YTMainVideoEntity = YTLocalEntity<EntityType.mainVideoEntity>['data'] & {
+type YTMainVideoEntity = Partial<YTLocalEntity<EntityType.mainVideoEntity>['data']> & {
   addedTimestamp: number
   isAutoDownload: boolean
 }
 
+function copyEventTarget(event: Event): void {
+  const { target } = event
+
+  if (target instanceof HTMLElement) navigator.clipboard?.writeText(target.innerText)
+}
+
 const YTMainVideoEntityTableItem = (
+  filter: string,
   onWatch: (id: string) => void,
   onExport: (id: string, format: ExportFormat) => void,
   onDelete: (id: string) => void,
-  { title, videoId, addedTimestamp, isAutoDownload }: YTMainVideoEntity
-): ChildDom => {
-  return tr(
-    td(SINGLE_LINE_COLUMN, `${isAutoDownload ? '[Auto] ' : ''}${videoId ?? '-'}`),
-    td({ style: 'cursor:pointer;text-decoration:underline', onclick: onWatch.bind(null, videoId) }, title ?? '-'),
-    td(SINGLE_LINE_COLUMN, new Date(addedTimestamp).toLocaleString()),
-    td(
-      SINGLE_LINE_COLUMN,
-      button({ onclick: onExport.bind(null, videoId, ExportFormat.BUNDLE) }, 'Bundle'),
-      button({ onclick: onExport.bind(null, videoId, ExportFormat.AUDIO_STREAM) }, 'Audio'),
-      button({ onclick: onExport.bind(null, videoId, ExportFormat.VIDEO_STREAM) }, 'Video')
+  { owner, title, videoId, addedTimestamp, isAutoDownload }: YTMainVideoEntity
+): ChildDom[] => {
+  const type = isAutoDownload ? 'Auto' : 'Manual'
+  const timestamp = new Date(addedTimestamp).toLocaleString()
+  const search = [owner, title, videoId, type, timestamp]
+
+  if (videoId == null || (filter && !search.some(part => `"${part}"`.toLowerCase().includes(filter)))) return []
+
+  return [
+    tr(
+      td(COLUMN_PROPS_INFO_TITLE, title ?? '-'),
+      td(
+        COLUMN_PROPS_ACTION_BODY,
+        button({ onclick: onExport.bind(null, videoId, ExportFormat.BUNDLE) }, 'Bundle'),
+        button({ onclick: onExport.bind(null, videoId, ExportFormat.AUDIO_STREAM) }, 'Audio'),
+        button({ onclick: onExport.bind(null, videoId, ExportFormat.VIDEO_STREAM) }, 'Video')
+      ),
+      td(
+        COLUMN_PROPS_ACTION_BODY,
+        button({ onclick: onWatch.bind(null, videoId) }, 'Watch'),
+        button({ onclick: onDelete.bind(null, videoId) }, 'Delete')
+      )
     ),
-    td(SINGLE_LINE_COLUMN, button({ onclick: onDelete.bind(null, videoId) }, 'Delete'))
-  )
+    tr(
+      td(COLUMN_PROPS_INFO_SMALL, videoId ?? '-'),
+      td(COLUMN_PROPS_INFO_SMALL, owner ?? '-'),
+      td(COLUMN_PROPS_INFO_SMALL, `[${type}] ${timestamp}`)
+    )
+  ]
 }
 
 class YTOfflinePageLifecycle extends Lifecycle<void> {
@@ -58,6 +87,8 @@ class YTOfflinePageLifecycle extends Lifecycle<void> {
   private readonly mainVideoEntities_: State<YTMainVideoEntity[] | null>
   private readonly password_: State<string>
   private readonly status_: State<string>
+  private readonly queueVideoId_: State<string>
+  private readonly filter_: State<string>
   private refreshTimer_: ReturnType<typeof setInterval> | null
 
   public constructor() {
@@ -67,11 +98,13 @@ class YTOfflinePageLifecycle extends Lifecycle<void> {
     this.mainVideoEntities_ = van.state(null)
     this.password_ = van.state('')
     this.status_ = van.state('-')
+    this.queueVideoId_ = van.state('')
+    this.filter_ = van.state('')
     this.refreshTimer_ = null
   }
 
   protected override onCreate(): void {
-    const { classList, fileInput_, mainVideoEntities_, password_, status_ } = this
+    const { classList, fileInput_, mainVideoEntities_, password_, status_, queueVideoId_, filter_ } = this
 
     const className = buildClass(['bu-overlay', 'page'])
     classList.add(className)
@@ -87,6 +120,7 @@ class YTOfflinePageLifecycle extends Lifecycle<void> {
     const refreshTable = (): void => {
       getYTLocalEntitiesByType(EntityType.mainVideoEntity, true)
         .then(entities => Promise.all(entities.map(async ({ data }) => {
+          const channel = await getYTLocalEntityByKey<EntityType.ytMainChannelEntity>(data.owner, true)
           const downloadContext = await getYTLocalEntityByKey<EntityType.videoDownloadContextEntity>(encodeEntityKey({
             entityId: data.videoId,
             entityType: EntityType.videoDownloadContextEntity,
@@ -100,16 +134,15 @@ class YTOfflinePageLifecycle extends Lifecycle<void> {
 
           return {
             ...data,
+            owner: channel?.data.title,
             addedTimestamp: downloadState?.data.downloadStatusEntity.downloadState === 'DOWNLOAD_STATE_COMPLETE' ? Number(downloadState.data.addedTimestampMillis) : -1,
             isAutoDownload: downloadContext?.data.offlineModeType === 'OFFLINE_MODE_TYPE_AUTO_OFFLINE'
           }
         })))
         .then(entities => {
-          mainVideoEntities_.val = entities.filter(entity => entity.addedTimestamp >= 0).sort((l, r) => (
-            l.isAutoDownload === r.isAutoDownload ?
-              (r.addedTimestamp - l.addedTimestamp) :
-              (l.isAutoDownload ? 1 : -1) // NOSONAR
-          ))
+          mainVideoEntities_.val = entities
+            .filter(entity => entity.addedTimestamp >= 0)
+            .sort((l, r) => (r.addedTimestamp - l.addedTimestamp) * ((Number(l.isAutoDownload) - Number(r.isAutoDownload)) || 1))
         })
         .catch(error => status_.val = error instanceof Error ? error.message : String(error))
     }
@@ -162,7 +195,27 @@ class YTOfflinePageLifecycle extends Lifecycle<void> {
     }
 
     const handleDelete = (id: string): void => {
+      if (!confirm(`Confirm delete video '${id}'?`)) return
+
       promiseStatus(deleteYTOfflineMedia(id)).finally(refreshTable)
+    }
+
+    const handleQueueDownload = (): void => {
+      const videoId = queueVideoId_.val.trim()
+      if (videoId.length === 0) return
+
+      dispatchYTAction({
+        actionName: 'yt-offline-video-endpoint',
+        args: [{
+          offlineVideoEndpoint: {
+            action: 'ACTION_ADD',
+            actionParams: { formatType: 'HD_1080', settingsAction: 'DOWNLOAD_QUALITY_SETTINGS_ACTION_DONT_SAVE' },
+            videoId
+          }
+        }],
+        returnValue: []
+      })
+      queueVideoId_.val = ''
     }
 
     this.refreshTimer_ = setInterval(refreshTable, 60e3)
@@ -172,21 +225,44 @@ class YTOfflinePageLifecycle extends Lifecycle<void> {
 
     van.add(
       this,
-      h1('Downloaded Videos'),
+      h1('Import/Export'),
       div(
-        { style: 'display:flex;flex-direction:row;align-items:center;gap:0.5em' },
-        button({ onclick: refreshTable }, 'Refresh'),
-        button({ onclick: handleImport }, 'Import Bundle'),
-        input({ style: FLEX_ITEM_GROW_STYLE, placeholder: 'Bundle password (Optional)', value: password_, oninput: e => password_.val = e.target.value }),
-        p({ style: FLEX_ITEM_GROW_STYLE }, () => `Status: ${status_.val}`)
+        { style: `${STYLE_FLEX_CONTAINER};flex-basis:50%` },
+        div(
+          { className, style: STYLE_FLEX_ITEM_GROW },
+          h4('Import Video'),
+          div(
+            { style: STYLE_FLEX_CONTAINER },
+            button({ onclick: handleImport }, 'Import Bundle')
+          ),
+          p(() => `Status: ${status_.val}`)
+        ),
+        div(
+          { className, style: STYLE_FLEX_ITEM_GROW },
+          h4('Options'),
+          input({ placeholder: 'Bundle Password (Optional)', value: password_, oninput: e => password_.val = e.target.value }),
+        )
       ),
+      h1('Download (Premium Only)'),
+      div(
+        { style: STYLE_FLEX_CONTAINER },
+        input({ placeholder: 'Video ID', value: queueVideoId_, oninput: e => queueVideoId_.val = e.target.value }),
+        button({ onclick: handleQueueDownload }, 'Queue Download')
+      ),
+      h1('Available Videos'),
+      input({ placeholder: 'Filter', value: filter_, oninput: e => filter_.val = e.target.value }),
       table(
-        thead(tr(th('ID'), th('Title'), th('Added On'), th('Export As'), th('Action'))),
+        thead(tr(
+          th(COLUMN_PROPS_INFO_HEAD, 'Title / ID / Channel / [Type] Added On'),
+          th(COLUMN_PROPS_ACTION_HEAD, 'Export As'),
+          th(COLUMN_PROPS_ACTION_HEAD, button({ onclick: refreshTable }, 'Refresh'))
+        )),
         () => {
           const entities = mainVideoEntities_.val
+          const filter = filter_.val.trim().toLowerCase()
           return tbody(
             entities?.length ?
-              entities.map(YTMainVideoEntityTableItem.bind(null, handleWatch, handleExport, handleDelete)) :
+              entities.flatMap(YTMainVideoEntityTableItem.bind(null, filter, handleWatch, handleExport, handleDelete)) :
               tr(td({ colSpan: 5 }, entities == null ? 'Loading...' : 'Videos you download will appear here')) // NOSONAR
           )
         }
